C0.tmLanguage
Added scopes for functions, structs & contract annotations, and then removed a scope which highlighted my name because I want to believe I don't have an ego problem. The regex for directives at the top I copied from the coin interpreter's help description, but the only one people would ever actually have is the #use. I can't remember if the original had broken string literals or if I introduced that error, but it works as expected now. I also couldn't figure out how to limit things by a number because the usual regex of (scope){repetitions}, so for hexadecimal literals, I copypasted the A-F0-9 match eight times. It's really sloppy.