Update C0.tmLanguage
Completely generic types, added fancy things for struct operations, function arguments, and variable declarations. Also, now it doesn't break the highlighting to do else if (), because the else capture comes before the function definition capture, whereas before it thought else was the type and if was the function name. Also fixed the hex literal capture, now it uses curly braces to indicate the number of repetitions. The function arguments are poorly delimited, but if the syntax is correct, they're ok. I would add an end clause for the newline character if function declarations always fit onto one line, as it is, I don't know how to fix it.